changed all calculations to decimal type
to stop rounding errors

+++ Foam_Calculator.Tests/FoamUnitPriceServiceTests.cs
@@ -22,11 +22,11 @@ namespace Foam_Calculator.Tests
 
 
             //act
-            double green50mmUnitPrice = foamPriceService.GetUnitPriceByColourAndThickness("Green", 50);
+            decimal green50mmUnitPrice = foamPriceService.GetUnitPriceByColourAndThickness("Green", 50);
             int green50mmSKU = foamPriceService.GetSkuByColourAndThickness("Green", 50);
 
             //assert
-            green50mmUnitPrice.Should().Be(0.9);
+            green50mmUnitPrice.Should().Be(0.9m);
             green50mmSKU.Should().Be(25085);
         }
